pyproject.toml
[tool.poetry]
name = "stock-linebot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
line-bot-sdk = "^3.14.2"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.1"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
